requirements.txt
goose3==3.1.16
matplotlib==3.7.1
numpy==1.23.5
pandas==2.0.2
plotly==5.15.0
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.23.1
vaderSentiment==3.3.2
vaderSentiment==3.3.2
